Cpu/mps fix for ray.remote
A little messy, but with respect to ray.remote(), the code is now optimised for both cpu and mps use – specifically, the code uses cpu for all predictions, and mps for gradient backprop and optimisation. Because ray.remote() can't handle devices on mps, this involves (for now) the creation of a second 'prediction net', which remains exclusively on the cpu.